using original add_torrent_params.hpp

+++ src/main/java/org/libtorrent4j/swig2/sha1_hash.java
@@ -36,7 +36,7 @@ public class sha1_hash {
     }
   }
 
-  public static int size() {
+  public static long size() {
     return libtorrent_jni.sha1_hash_size();
   }
